refactor(ui-mobileapp): switch relative unit size to px
