+++ ui-mobileapp/src/components/atoms/Card.ts
@@ -9,7 +9,7 @@ export type CardProps = {
 // NOTE: Pressable use here instead of View because the card is an interactive button
 export const Card = styled.Pressable<CardProps>`
   background-color: ${({ theme }) => theme.colors.white[500]};
-  padding: 1.25rem;
+  padding: 20px;
   border-radius: ${({ theme, rounded }) =>
     rounded ? theme.border.radius[rounded] : theme.border.radius.md};
 `;
