+++ ui-mobileapp/src/components/layouts/MainLayout.tsx
@@ -8,7 +8,7 @@ export default function MainLayout({ children }: MainLayoutProps) {
 }
 
 const Container = styled.View`
-  padding: 1.5rem;
+  padding: 24px;
   background-color: #fcfcfc;
   min-height: 100%;
 `;
