+++ ui-mobileapp/src/components/molecules/ShopCard.tsx
@@ -68,7 +68,7 @@ const StyledButton = styled(Button)`
 const InfoWrapper = styled.View`
   position: relative;
   width: 100%;
-  height: 10rem;
+  height: 160px;
   // FIXME
   border-top-left-radius: ${({ theme }) => theme.border.radius.lg};
   border-top-right-radius: ${({ theme }) => theme.border.radius.lg};
