+++ ui-mobileapp/src/components/organisms/LoginForm/LoginForm.tsx
@@ -82,5 +82,5 @@ export default function LoginForm() {
 
 const InputWrapper = styled.View`
   display: grid;
-  grid-gap: 0.5rem;
+  grid-gap: 8px;
 `;
