+++ ui-mobileapp/src/components/organisms/RegisterSteps/RegisterStepTwo.tsx
@@ -120,5 +120,5 @@ export default function RegisterStepTwo({ nextStep }: RegisterStepTwoProps) {
 
 const InputWrapper = styled.View`
   display: grid;
-  grid-gap: 0.5rem;
+  grid-gap: 8px;
 `;
